plantai-backend/requirements.txt
fastapi==0.128.8
uvicorn==0.39.0
httpx==0.28.1
pydantic==2.12.5
shapely==2.0.7
python-dotenv
